vector: refactor code to remove a global instance.
This kind of global static instance is really dangerous in c++ because
We have no idea when the instance will be initalization exactly.

This might interrupt the program sequence, more worsely, it could break the program
while That breakage is totally up to the member instances of the class.

Very poor design.